requirements.txt
bcrypt==4.1.2
blinker==1.7.0
click==8.1.7
Flask==3.0.2
Flask-Bcrypt==1.0.1
Flask-CKEditor==0.5.1
Flask-Cors==4.0.0
Flask-Uploads==0.2.1
Flask-WTF==1.2.1
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
psycopg2-binary==2.9.9
shortuuid==1.0.13
Werkzeug==3.0.1
WTForms==3.1.2
